+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 Changelog
 =========
 
+2.0 (unreleased)
+----------------
+
  * **2016-05-08**: [BC BREAK] Removed `showAtion` in favor of `listAction` in the `SuggestionProviderController`
  * **2016-05-02**: [BC BREAK] Dropped PHP <5.5 support
  * **2016-05-02**: [BC BREAK] Dropped Symfony <2.8 support
+
+1.3.0
+-----
+
+* **2016-06-09**: Build and register LastModificationGuesser for PHPCR to use a last modification date on a sitemap. 
+* **2016-04-12**: Build and register DepthGuesser for PHPCR to use depths information for structure sitemap 
 
 1.2.0
 -----